removed unused elements
