--- Sombra.TemplateService/Templates/ChangePasswordTemplate.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Sombra.TemplateService.Templates
-{
-    class ChangePasswordTemplate 
-    {
-    }
-}
